split logic
